Debug egl changes on Windows. Move egl helpers into xr.utils.gl.egl_util Move opengl helpers into xr.utils.gl Move PySide6OffscreenContextProvider into xr.utils.gl.pyside Move glfw helpers into xr.utils.gl.glfw_util
