+++ src/xr/utils/__init__.py
@@ -2,17 +2,14 @@ from . import classes
 from . import context_object
 from . import graphics_context_provider
 from . import matrix4x4f
-from . import opengl_graphics
 
 from .classes import *
 from .context_object import *
 from .graphics_context_provider import *
 from .matrix4x4f import *
-from .opengl_graphics import *
 
 __all__ = []
 __all__.extend(classes.__all__)
 __all__.extend(context_object.__all__)
 __all__.extend(graphics_context_provider.__all__)
 __all__.extend(matrix4x4f.__all__)
-__all__.extend(opengl_graphics.__all__)
